#9
Fixed bug in which classes not on the classpath of the org.gamegineer.common.persistence bundle could not be deserialized even though they had a registered persistence delegate.  It turns out the existing adapter manager based method for registering persistence delegates is not robust enough.  That solution requires an existing object in order to retrieve the associated persistence delegate.  However, upon deserialization, the first thing that must be retrieved is the Class object corresponding to the class name stored in the stream.  This responsibility has now been offloaded to the persistence delegate because it has access to the correct classpath for retrieving class being deserialized.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@1973 897c5806-f142-0410-a293-e1bf5698dec8