#10
Partial fix for the JavaBeans persistence framework class loading issue.
Replaced the use of IAdapterManager with a new registry specific for JavaBeans persistence delegates, similar to what was done for the Serializable persistence delegates.
If an object being serialized has a persistence delegate registered with the framework, the persistence delegate's class loader will be set as the context class loader while the object is being serialized.  This fix is only partial because the same class loader must be active during deserialization.  This will be addressed in a forthcoming change.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@1975 897c5806-f142-0410-a293-e1bf5698dec8