#10
Fixed JavaBeans persistence framework so it uses the correct classpath when serializing and deserializing objects.  The implementation supports any class that has a persistence delegate registered with the IPersistenceDelegateRegistry service.

During serialization, if a persistence delegate is registered for a class, a special object (ClassLoaderContext) is instantiated and written to the stream.  The instantiation of ClassLoaderContext activates the class loader for the object.  The previously active class loader is restored once the object has been written to the stream.

During deserialization, ClassLoaderContext instances are read transparently with the same effect as above regarding the activation of the class loader associated with the class following the ClassLoaderContext in the stream.  Again, the previously active class loader is restored once the object has been read from the stream.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@1976 897c5806-f142-0410-a293-e1bf5698dec8