#15
The keyboard accelerators stopped working because no component had the focus after opening the new table.  When no component has the keyboard focus, keyboard accelerators are not processed.  The root cause is due to how we destroy and recreate the TableView component when a new table is opened.  TableView is the component with the keyboard focus.  When it is destroyed, the keyboard focus is lost.  When the new instance is created, it does not by default get the keyboard focus.

The fix is to request the keyboard focus for the new TableView when it is created.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@1987 897c5806-f142-0410-a293-e1bf5698dec8