#22
Fixed race condition between the activation of the DebugOptions service and the ILoggingService service.  Instead of using Services to acquire a reference to each of these services in a non-deterministic manner, we now acquire both services on-demand.  Thus, instead of immediately acquiring them when the org.gamegineer.common.core bundle is activated, they are now acquired only when the org.gamegineer.common.core.runtime.Debug and org.gamegineer.common.core.runtime.Loggers classes are used.

The service references are no longer cached but rather re-acquired on every request.  If this turns out to be a performance issue, we can introduce ServiceTrackers in the Activator class.  I chose not to do that at this time to enforce separation of concerns because in both cases the acquisition of the services are implementation details of the Debug and Loggers classes.  However, this now introduces a dependency on OSGi in both of these classes, which is not that big of a deal because neither of these classes would likely every be used as POJOs.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2059 897c5806-f142-0410-a293-e1bf5698dec8