#23
Added support for dragging multiple cards from one card pile to another.  Currently this feature is only supported for a source card pile in any of the accordian layouts.  Had to add the ICardPile removeCards(Point), getBaseLocation(), and setBaseLocation(Point) methods.

Note that ICardPile.removeCards(Point) was not strictly necessary from a functionality perspective because the same thing could have been performed using existing methods on the ICardPile interface.  However, anticipating the need for atomic operations when we add support for networked players, this method was added so the entire operation could be performed in a single method call.

To use this feature, simply click on a visible card and begin a drag operation.  The mobile card pile will inherit the layout of the source card pile for the duration of the drag operation.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2096 897c5806-f142-0410-a293-e1bf5698dec8