#32
Automatically add an extension to the selected file in the Save As dialog if no extension is present.  The added extension is based on the active file filter.  If the All Files file filter is selected, no extension will be added.  If the active file filter has multiple extensions, the first extension will be used.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2149 897c5806-f142-0410-a293-e1bf5698dec8