#34
Fixed bug in the file history preferences that prevented it from distinguishing between equivalent files based solely on case.  Previously, the file history was maintained as a collection of Strings that represented the absolute path of each file.  On platforms where file names are case-insensitive, there was the possibility of specifying the same path multiple ways that only differed by case (e.g. the user typing in a path versus using the JFileChooser to navigate to the file).

Now the file history uses Files to track the entries.  I also refactored a lot of the methods that accepted file names throughout the model and changed them to now accept Files.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2150 897c5806-f142-0410-a293-e1bf5698dec8