#41
Fixed race condition that caused Swing data binding properties to fail because no SwingRealm was installed on the system event dispatch thread.  It is necessary to install the system SwingRealm after at least one AWT component has been created to ensure the EDT is not shut down, which would result in the loss of the previously installed SwingRealm.  The MainFrame class was chosen to serve as the life cycle manager for the system SwingRealm.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2237 897c5806-f142-0410-a293-e1bf5698dec8