Another major refactoring of the network table implementation.
Introduced network table strategies so that client and server network tables can be implemented separately.  Clients and servers have different resource and behavioral requirements, such that a single network table implementation is insufficient.

Introduced the table gateway as an intermediate abstraction to act as a broker between the network table and a local or remote table.  Remote gateways are also responsible for providing a network service and handling all messages appropriately.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2343 897c5806-f142-0410-a293-e1bf5698dec8