Modified transport layer to propagate exceptions all the way up to the network table components so the user can be properly informed of the reason for a passive disconnect.
For example, if the server process is terminated before it can send a FIN, the client will eventually experience an I/O error.  Without being able to propagate this exception through the transport interfaces, the client's user will never receive a notification that the network was disconnected in a non-normal manner.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2372 897c5806-f142-0410-a293-e1bf5698dec8