Refactored authentication algorithm to set up a future change whereby the network table strategy will send a PlayerList message to the newly-connected gateway from tableGatewayAdded().  The PlayerList message must be sent after the EndAuthentication message.  Therefore, the addTableGateway() method must be called after authentication is successful.
Related to above, we must ensure that the network table remains locked during the entire execution of the IMessageHandler.handleMessage() method.  This allows the message handler to implement algorithms that would normally be affected by TOCTOU conditions.  All methods on ITableGatewayContext now require the caller to acquire the context instance lock before calling the method.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2380 897c5806-f142-0410-a293-e1bf5698dec8