#46
Fixed deadlock within network transport layer.

The code in Dispatcher.close() to forcibly close orphaned event handlers should not be run while the Dispatcher is locked.  The code was modified to execute this block while the Dispatcher is unlocked (except for a small window required to safely copy the event handler collection).

Theoretically, this could introduce a race condition if another thread were to call Dispatcher.close() at the same time.  That is, the orphaned event handlers could be closed multiple times.  However, the ServiceHandler.close() method is thread-safe and gracefully handles calls to close() even if the handler is not in the OPEN state.  It also tracks tracks whether the handler is currently registered with the Dispatcher and guarantees the handler will only ever be unregistered once, thus avoiding any possible IllegalArgumentExceptions being thrown from Dispatcher.unregisterEventHandler().

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2442 897c5806-f142-0410-a293-e1bf5698dec8