Fixed another latent deadlock in the node layer.  The deadlock became apparent now that the server is sending a GoodbyeMessage to connected clients when it is shut down.  This causes a near simultaneous passive and active disconnect of the client.
The first is the original disconnect issued by the user to shut down the server.  While shutting down the transport layer, the following locks are acquired:

  LocalNode -> (via transport layer) RemoteNode -> LocalNode (reentrant)
  
The second is the passive disconnect seen after the remote client closes the socket.  When the transport layer initiates this disconnect, the following locks are acquired:

  RemoteNode -> LocalNode

Because the LocalNode and RemoteNode locks are taken in a different order, a deadlock results.

The solution is not to hold the LocalNode lock when calling into the transport layer.  In order to ensure atomicity of the active disconnect request by the user, a new lock was introduced to serialize access to the LocalNode through its controller interface (i.e. the connect() and disconnect() methods invoked by the TableNetwork object).  The instance lock for LocalNode is only held when needed during disconnect() to ensure memory consistency.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2445 897c5806-f142-0410-a293-e1bf5698dec8