Removed the boolean return value from IServiceContext.sendMessage().  It really cannot be used to indicate if the message was sent successfully.  The only possible error that can occur is a programmer error due to not implementing a persistence delegate for the message being sent.
This will ultimately help us to further decouple the node layer from the transport layer.  Currently, there is the possibility for a deadlock due to the locks for remote nodes and their corresponding service contexts being taken in a different order along different code paths.  In the future, we will most likely decouple these two layers using a lock-free queue to transfer requests.  In such an implementation, it would be impossible to return a meaningful value from sendMessage() anyway due to its asynchronous nature.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2481 897c5806-f142-0410-a293-e1bf5698dec8