Implemented optimization to improve rendering of card movements on remote tables.
In the current implementation, handling the ICardListener.cardLocationChanged() event is unnecessary because a card can only be moved by moving a card pile that contains the card. Thus, when the corresponding ICardPileListener.cardPileBoundsChanged() event is fired, it will automatically set the new card location when the card pile location is changed.

If we send both the card pile bounds changed event and the card location changed event over the network, the card movement on remote tables will appear "jumpy" because two ICard.setLocation() calls will be made that may be a few pixels off.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2491 897c5806-f142-0410-a293-e1bf5698dec8