Modified MainModel and TableModel to behave more like an SDI application.  That is, instead of creating a new TableModel every time the user opens a new or existing table, the existing table model is reused.
This was done primarily to make the New and Open commands more network friendly.  Previously, we had disabled these commands while the network was connected because it destroyed the existing table and table network.  Now, these commands can freely be used.

If we every move to an MDI architecture, the New and Open commands will once again create a new TableModel, but it will appear in a separate window, and thus not affect any existing table or network connections.

In addition, the Import command was removed from the Table menu because it is now superfluous.  If we ever move to an MDI architecture, it will have to be added once again.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2498 897c5806-f142-0410-a293-e1bf5698dec8