Fixed test break.
A class circularity error was introduced when we manually registered the executor service in the common.core bundle.  I thought the root cause of this issue was the fact that the common.core Activator was referencing the Debug class during bundle shutdown.  However, this turned out not to be the case.

The root cause was related to the manual registration of the ExecutorService in the common.core bundle's start() method.  Not quite sure why there is a dependency between this manually registered service and the other lazily declared services.  Changing the executor service to a declarative service fixed the problem.  It also addressed a TODO regarding not creating the underlying ExecutorService until it was actually needed by another bundle in the system.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2553 897c5806-f142-0410-a293-e1bf5698dec8