Fixed a potential thread safety issue in the TCP transport layer.
The TCP transport layer uses thread confinement to ensure thread safety instead of locks.  However, the transport layer constructor was being called on a thread other than the transport layer thread.  This had the potential of causing the non-final fields to not reflect the correct value on the transport layer thread.

Modified the TcpTransportLayerFactory class to use an intermediate factory whose responsibility is to first create the transport layer executor service, and then to instantiate the transport layer itself on the transport layer thread.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2566 897c5806-f142-0410-a293-e1bf5698dec8