Fix for regression that prevented authentication errors from being correctly displayed.
The root cause is due to the following sequence of events.  An ErrorMessage is received containing the authentication error and is stored properly in the call to AbstractRemoteNode.close(TableNetworkError).  The server immediately closes the socket, which causes the client's input stream to return EOF, which causes the ServiceHandler to invoke AbstractRemoteNode.peerStopped().  peerStopped() invokes AbstractRemoteNode.close(TableNetworkError.UNEXPECTED_PEER_TERMINATION), and the original error is overwritten.

The fix is to guard against overwriting the error on repeated calls to AbstractRemoteNode.close(TableNetworkError).  Similar to what is done in AbstractRemoteNode.stopped(Exception).

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2579 897c5806-f142-0410-a293-e1bf5698dec8