#62
Fixed the case of calling InputQueue.dequeueMessageEnvelope() with a partial message envelope in the queue.  The fix is to record the buffer position and limit before calling flip() so they can be restored in the event a complete message envelope is not available.

Also added a new unit test to cover the case when more than one message envelope is in the queue.  This test ensures the branch that calls ByteBuffer.compact() is covered by a test.  I wasn't sure if anything else had to be done in this case to possibly undo a side effect of the flip() call, but apparently everything is working fine.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2616 897c5806-f142-0410-a293-e1bf5698dec8