#58
Ensure that transport layer messages larger than buffer pool size (currently 16 kB) are not truncated.  Fixed by replacing the single buffer in InputQueue with an expandable collection of buffers.

Had to hack up the MessageEnvelope class a bit to allow it to handle collections of ByteBuffers instead of a single ByteBuffer.  This class will be refactored in a future change to clean up this smell.

Also removed several unused methods from InputQueue.  While they make InputQueue more generic, they are not needed.  InputQueue is specifically written to deal with MessageEnvelopes, not generic byte streams.  Will probably clean up OutputQueue in a similar manner in the near future.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2617 897c5806-f142-0410-a293-e1bf5698dec8