#60
The root cause of this defect was due to a combination of two issues.

First, every state change in the table causes the Actions.updateAll() method to be invoked.  This is especially notorious when a card pile is being moved because not only does the card pile fire a locationChanged event, but all cards in the card pile also fire individual locationChanged events.  (This is apparently a design defect that will be corrected in the future.)  For example, moving a 52-card card pile will generate 53 locationChanged events per mouse move event, which will cause 53 calls to Actions.updateAll().

Second, many action enablement callbacks invoke ITableNetwork.getLocalPlayer() to determine if the local network player has the editor role.  When the network is connected, this call is dispatched to the node layer thread while the caller waits.

The combination of the large number of calls to Actions.updateAll() and the fact that each of these calls may invoke ITableNetwork.getLocalPlayer() multiple times is the root cause of the performance problem observed when moving a large card pile while the network is connected.

The fix that was implemented is to throttle the calls to Actions.updateAll().  Instead of invoking this method each time a table state change event is received, a flag is set that indicates the actions should be updated.  A separate thread periodically wakes up and invokes Actions.updateAll() if the flag is set.  Currently, this thread wakes up every 250 ms to ensure keyboard shortcuts appear responsive to the state of the application.

In order to ensure menu items themselves are enabled as quickly as possible, all menus invoke Actions.updateAll() immediately before they are displayed.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2635 897c5806-f142-0410-a293-e1bf5698dec8