#61
Added the IHelpSystem.shutdown() method to overcome a quirk in the JavaHelp system.

We dynamically add and remove help sets as bundles come and go.  There is a "feature" of the Search view in which it issues a AWT beep() after running a search.  A search happens to be run when a help set is removed from the master help set.  This always leads to an annoying beep when shutting down the application if the help window was ever displayed during the session.

The workaround is to ignore requests to add and remove help sets to the master help set after the new shutdown() method has been called.  The application invokes shutdown() after it has been instructed to close, which typically occurs before the framework begins shutting down bundles (including the bundles that contribute help set providers).

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2684 897c5806-f142-0410-a293-e1bf5698dec8