#63
The root cause of this defect is due to how the CardPileView.dirtyBounds_ field is managed.  The purpose of this field is to keep track of the previous bounds of the card pile so that when a card pile bounds changed event is received, we can force a repaint of the union of the old card pile bounds and the new card pile bounds.

This field accumulated the new card pile bounds upon reception of the card pile bounds changed event, which is correct.  However, it was reset at the conclusion of the paint() method.  It was assumed that the only event that could trigger a call to paint() was a subsequent call to TableView.repaintTable() (which called TableView.repaint()).

However, there are other events in the system that can cause a repaint.  In such cases, the dirtyBounds_ field was prematurely reset to the current bounds, which opened the possibility that the old card pile bounds would never be invalidated during the call to TableView.repaintTable().

The fix is to only reset the dirtyBounds_ field only after a call to TableView.repaintTable().  To that end, the CardPileView.cardPileBoundsChanged() method was modified to create the union of the old and new card pile bounds, invalidate that union, and then reset dirtyBounds_ to the new card pile bounds.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2690 897c5806-f142-0410-a293-e1bf5698dec8