#64
Fixed race condition in TableRunner.call().  Modified methods that asynchronously open and close the frame window on the Swing EDT to now wait for those operations to complete before continuing.  This prevents the TableRunner.call() method of racing ahead of the Swing EDT thread and potentially shutting down the OSGi framework while code still runs on the EDT.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2697 897c5806-f142-0410-a293-e1bf5698dec8