Added support for persisting the table's root component in the table memento.
Reverted [2436] to reintroduce the MementoUtils.getRequiredAttribute() and getOptionalAttribute() methods.  Optional attributes are necessary once again now that the table's root component is optional.

Fixed buggy implementations of createMementoOriginator() in AbstractComponentTestCase and AbstractTableTestCase.  This method MUST return a new instance.  If it returns the same instance, the test will always pass even if the memento save/restore functionality is broken.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2824 897c5806-f142-0410-a293-e1bf5698dec8