Removed FIXME.  We will continue to programatically register the NULL component surface design as it is the ultimate fallback surface design.
The information in the comment is still applicable.  Namely, if we don't register built-in component surface designs directly in the component surface design registry, they are not guaranteed to be registered by the time bundles at the default start level are started.  Declarative registrations happen before the default start level because we've configured the application to start the DS bundle earlier.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2915 897c5806-f142-0410-a293-e1bf5698dec8