#68
Fixed bug in TableView.repaintTable().  The Rectangle input to this method was being modified by the method as a side effect ONLY WHEN THE VIEW ORIGIN OFFSET WAS NON-ZERO.  In all cases, the source Rectangle passed to this method was ComponentView.dirtyBounds_.

In the case where the view origin offset was non-zero, the dirty bounds were being unexpectedly modified.  The dirty bounds are always supposed to be reinitialized to the current bounds after a repaint.  However, this information was being lost, and thus the original location of a dragged component was not included in the repaint operation.

The fix is to create a copy of the input Rectangle before potentially modifying it.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2918 897c5806-f142-0410-a293-e1bf5698dec8