Fixed a race condition in several places that ultimately caused an assertion failure in the UI.
In several places, we were performing the following sequence of operations:

  1) add a container listener to a container
  2) get the list of child components from the container

The race condition was that in between (1) and (2), one or more components could have been added to the container.  In such a case, some code would process the list of child components obtained in (2) which includes the components added after (1).  Then, at a later time, one componentAdded event was delivered for each component added between (1) and (2).

In the ContainerView class, this was fatal because the view assumed it would only ever register a child view for each child component once.  However, due to the race condition, this could occur twice resulting in an assertion failure in ContainerView.createComponentView().

The fix is to ensure the registration of the container listener and obtaining the list of child components occurs atomically.  A new helper method was added in the view to do this.

The same issue occurred in the network code.  Here, the fix is to ensure the table environment is locked during the calls to LocalNetworkTable.addComponentListeners() and removeComponentListeners().

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2922 897c5806-f142-0410-a293-e1bf5698dec8