Modified how BundleSuiteBuilder searches for test classes in order to no longer require test fragments to package their classes in a JAR file.
We now first assume the test is being run from the IDE and search the IDE classpath as before.  If no classes are found using that method, we fall back to assuming the test is being run from a deployment and search the bundle classpath.

If the bundle classpath is not explicitly specified, we assume it is the root of the bundle (".") rather than nothing.  This matches the default behavior according to the OSGi specification.

We also handle both JAR files and folders as classpath entities.  Previously, we assumed the IDE classpath only consisted of folders and the bundle classpath only consisted of JAR files.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2931 897c5806-f142-0410-a293-e1bf5698dec8