#69
Fixed race condition in which the order of delivery of table component events to the UI view and model was not deterministic.

Redesigned the views to no longer directly listen to their associated table component.  Rather, all events are now routed through their corresponding model.  This ensures the model always receives the event before the view, so that the view can always assume the model reflects the latest state of the table at the time the event was delivered.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2933 897c5806-f142-0410-a293-e1bf5698dec8