#70
Fixed issue when the focused component is deleted, but the TableModel still reports that component has the focus.  The root cause was due to the ContainerModel.ContainerListener.componentRemoved() method calling setFocused(false) on the component being removed.  While this would correctly fire a componentModelFocusChanged event up the listener chain, the TableModel never changed the value of focusedComponent_.

The fix is for the componentRemoved() method to call TableModel.setFocus(null) if the component being removed had the focus.  This required injecting a reference to the TableModel into the ComponentModel hierarchy.

This solution has a few bad smells that will be addressed outside the scope of this ticket.  A new unit test was added to verify we do not introduce a regression while refactoring the fix to this ticket.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2939 897c5806-f142-0410-a293-e1bf5698dec8