Changed the componentPrototypes extension point by extracting the factory class name attribute to a child element in order to conform with the executable extension requirements defined by IConfigurationElement.createExecutableExtension().
Namely, in order to support a type 2 executable extension (one that is passed a collection of parameters as initialization data), the class must appear in a child element with an attribute named "class" that contains the fully-qualified name of the executable extension.  Also, the child "parameter" elements are now children of the new "factory" element.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2971 897c5806-f142-0410-a293-e1bf5698dec8