Added support for enablement expressions to the componentPrototypes extension point.
This required modifying the user interface model to support generation of evaluation contexts based on the current model state.  Currently, the only expression variable supported is the focused component on the table.

Also, the user interface model had to be redesigned to keep track of each component model's parent rather than just the table model that owns each component model.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@2981 897c5806-f142-0410-a293-e1bf5698dec8