Major redesign of how the user sets focus in the UI in order to handle the forthcoming feature that allows any component to be moved anywhere on the table instead of just containers.
This change addresses how a user sets the focus when there are multiple components underneath the cursor.  It is necessary for the user to select any of the focusable components under the cursor, not simply the top-most component.

The first time the user left- or right-clicks the mouse over a collection of components, the top-most component receives the focus.  Each subsequent time the user left-clicks the mouse over a collection of components and one of the components has the focus, the first focusable ancestor of the focused component will receive the focus.  If no ancestor can receive the focus, the top-most focusable component will receive the focus.  In this way, continually left-clicking will iterate over all focusable components under the mouse, eventually returning to the originally focused component.

This change can be thought of as moving the focus along the ancestor axis.  Future changes will address moving the focus along the descendant, previous-sibling, and following-sibling axes.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3024 897c5806-f142-0410-a293-e1bf5698dec8