Extracted logic for drawing hover/focus border from ComponentView to TableView so that the entire border can be drawn.  Drawing the border in ComponentView restricted it to only drawing the border around the visible portion of the component.  TableView has full access to the entire graphic context, so it can draw the complete border, including the portion around the obscured area of the component.  Drawing the complete border looks like a strange graphical artifact, but it is intentional because it gives the user the correct context as to what component has the hover/focus.
Based on the above change, most of the logic regarding component view padding (the space between a component's surface and its border) was removed.  A future change will remove the remaining bits so that only the TableView will know anything about padding.

Modified ComponentView to alpha blend the active surface image with the focus color to provide better feedback to the user regarding the current hovered component.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3025 897c5806-f142-0410-a293-e1bf5698dec8