#76
Allow deletion of arbitrary components.

The component with the focus will be deleted when executing the Remove action.

If the component with the focus is a container, then executing the Remove All action will delete all child components of the container.  If no component has the focus, the Remove All action will delete all components on the tabletop.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3075 897c5806-f142-0410-a293-e1bf5698dec8