Fixed regression caused by [3082].
Refreshing the container layout after a component is removed caused an unintended side effect during a drag-and-drop operation.  After each component is removed from the source container and added to the mobile container, the remaining components to be dragged would have their position changed due to the new layout.  This ultimately caused all drag components to have the same location in the mobile pile.

The fix is to reset the location of each drag component once they've been added to the mobile container to what it was at the start of the drag-and-drop operation.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3094 897c5806-f142-0410-a293-e1bf5698dec8