#79
Fixed issue with hover rectangle still being drawn after the ancestor of the hovered component is deleted.

The root cause of this problem was much more fundamental than some defect in the UI layer.  Within the table, when a component is deleted, we were not explicitly deleting the descendants in the case the deleted component is a container.  While the content of the table was still correct after the deletion, we were breaking a contract with table listeners in that they expect to be informed when EVERY component is deleted, regardless of whether the component is explicitly or implicitly deleted.

The fix is to invoke IContainer.removeAllComponents() on each container that is deleted.  This will ultimately be called recursively for each descendant container.  New unit tests were added to ensure the appropriate events are fired when components are implicitly deleted.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3112 897c5806-f142-0410-a293-e1bf5698dec8