#79
Fixed issue with hover rectangle still being drawn after the ancestor of the hovered component is deleted.

Previously, the UI layer would receive a component removal event and check to see if the removed component currently had the hover and/or the focus, and if so, cleared the hover and/or focus.  The fix is to also check if the current hover and/or focus component is a descendant of the removed component, and if so, to clear the hover and/or focus as before.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3117 897c5806-f142-0410-a293-e1bf5698dec8