#80
Fixed race condition in table event delivery.  The race condition occurred because the path of the component that fired the event may have changed since the event was fired.  Most receivers retrieved the path of the component associated with the event by simply calling event.getComponent[Container]().getPath().  This caused all sorts of problems if the component had, for example, been removed from the table (and thus its path was null) before any prior events (when it was still part of the table) were delivered.

The fix is to provide the component path at the time the event was fired as an attribute of the event.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3118 897c5806-f142-0410-a293-e1bf5698dec8