Experimental change that will be immediately reverted due to several problems.
The driver for this change was an observation that the current event delivery mechanism from a table environment does not ensure events are delivered in the same order in which they were fired by the underlying table environment components.  This is due to the fact that each thread maintains its own event notification queue.  If concurrent changes are made to a table environment from multiple threads, receivers may not receive those events in the same order they were fired.

This can have consequences for many listeners which assume the state of the table environment proceeds from one consistent state to another.  Out-of-order event delivery may cause runtime exceptions to be raised unexpectedly.  It also can lead to incorrect state for listeners who incrementally change their state based on what they expect are incremental changes in the state of the table environment.

This experimental change marshals all table environment event notifications to a separate thread where they will be delivered in the same order they were fired (NB: this change is missing a significant piece of this guarantee, namely that the various fireXXXEvent() methods within the table environment implementation should be executed while the table environment lock is held).

However, they are several bad smells related to this implementation.  Namely, the significant changes made to the unit tests due to the asynchronous nature of event delivery.  I'm committing it simply to capture the tremendous amount of work that went into this exercise.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3121 897c5806-f142-0410-a293-e1bf5698dec8