#82
Now that we've introduced a single coarse-grained lock for the TableEnvironmentModel and all of its components, and that lock is exposed for client use, we need to ensure it's not possible for a model object to call out to a listener while the lock is held.

We employ a design similar to what we used to do for the original TableEnvironment (before we had to ensure events were guaranteed to be delivered in the same order they were fired; that is not a requirement here).  Basically, if the lock is held, we queue event notifications on the current thread until the lock is released and then fire them.  There is a pending event queue for each thread that accesses the TableEnvironmentModel and its components.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3129 897c5806-f142-0410-a293-e1bf5698dec8