Fix for regression introduced in [3126] and only partially fixed by #82.
This change addresses the root cause of the race condition introduced by [3126] in which table environment events are delivered asynchronously.  Previously, we assumed the ITable and its associated TableModel were kept synchronized because table events were delivered synchronously.  Now that that is no longer the case, there is a delay in the lifetime between an IComponent and its associated ComponentModel (e.g. after a new IComponent is added to the ITable, the corresponding ComponentModel will be added to the corresponding TableModel a finite time later).

The root cause was that the UI components (namely TableView) would try to manipulate the IComponents and ComponentModels simultaneously always assuming they were synchronized.  We now only manipulate ComponentModels, and thus are guaranteed a consistent view (although it may slightly lag behind the ITable, the TableModel will eventually catch up, and thus the TableView will render it correctly).

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3130 897c5806-f142-0410-a293-e1bf5698dec8