Added the ITableEnvironmentModelLock interface.  The implementation simply delegates to the ITableEnvironmentLock for the ITableEnvironment associated with the TableEnvironmentModel.
Cleaned up scope of acquiring the table environment model lock throughout the model package to ensure all public operations are atomic.  Removed acquisition of the table environment model lock from all private and package-private members, instead requiring the invoking method to acquire the lock on their behalf.

Had to fire the tableModelFocusChanged and tableModelHoverChanged events from TableModel.setFocus() and TableModel.setHover(), respectively, because the same events fired by ComponentModel may not propagate up the hierarchy in the case when an intermediate ancestor is removed due to how we queue events until the table environment model lock is released on the current thread.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3131 897c5806-f142-0410-a293-e1bf5698dec8